Large Commit changes, read following paragraph:
Added several commands, such as moveToSwitch, moveToScale, moveToGround, and moveToClimb, all for autonomous movement of the elevator. Not only that, but intake, outtake, and runClimber were all added as well. In the Elevator Subsystem, a new function, checkError(), was created to efficiently check the error of the encoders on the elevator.